requirements.txt
prefect
minio
pandas
pyarrow
faker
streamlit
plotly
python-dotenv
pymongo
uvicorn